protect a rutas de productos

+++ backend/controllers/productControllers.js
@@ -4,7 +4,7 @@ const Product = require('../models/productModel')
 //para obtener la lista de productos
 const getProducts = asyncHandler( async (req, res) => {
 
-    const products = await Product.find()
+    const products = await Product.find({user: req.user.id})
 
 
     res.status(200).json(products)
@@ -19,36 +19,50 @@ const createProducts = asyncHandler( async  (req, res) => {
     }
 
     const product = await Product.create({
-        name: req.body.name
+        name: req.body.name,
+        user: req.user.id //para poder crear un producto con un id de usuario
     })
     res.status(201).json(product)
 })
 
 //para modificar productos
 const updateProducts = asyncHandler( async  (req, res) => {
-    const product = Product.findById(req.params.id)
+    const product = await Product.findById(req.params.id)
 
     if(!product) {
         res.status(400)
         throw new Error('ese producto no existe')
     }
 
-    const productUpdated = await Product.findByIdAndUpdate(req.params.id, req.body, {new: true})
-    res.status(200).json(productUpdated)
+    //nos aseguramos que la producto pertenezca al usuario logeado, es decir el del token
+    if(product.user.toString() !== req.user.id) { //req.user viene de la funcion protectora protect
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+        
+    } else {
+        const productUpdated = await Product.findByIdAndUpdate(req.params.id, req.body, {new: true})
+        res.status(200).json(productUpdated)
+    }
 })
 
 //para eliminar productos
 const deleteProducts = asyncHandler( async  (req, res) => {
-    const product = Product.findById(req.params.id)
+    const product = await Product.findById(req.params.id)
 
     if(!product) {
         res.status(400)
         throw new Error('ese producto no existe')
     }
 
-    await Product.deleteOne(product)
-
-    res.status(200).json({id: req.params.id})
+     //nos aseguramos que la producto pertenezca al usuario logeado, es decir el del token
+    if(product.user.toString() !== req.user.id) { //req.user viene de la funcion protectora protect
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+        
+    } else {
+        await Product.deleteOne(product)
+        res.status(200).json({id: req.params.id})
+    } 
 })
 
 module.exports = {
